+++ crates/prompt_gateway/src/lib.rs
@@ -2,6 +2,7 @@ use filter_context::FilterContext;
 use proxy_wasm::traits::*;
 use proxy_wasm::types::*;
 
+mod context;
 mod filter_context;
 mod hallucination;
 mod http_context;
